# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Atelier Commerce API",
    "version": "1.0.0",
    "description": "Single-currency BDT commerce API. Bearer authentication; cash on delivery; server-side ownership and transactional inventory."
  },
  "servers": [
    {
      "url": "http://localhost:8081"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "summary": "Create customer account",
        "operationId": "postApiAuthRegister",
        "tags": [
          "Auth"
        ],
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Auth"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Register"
              }
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Sign in",
        "operationId": "postApiAuthLogin",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Auth"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "summary": "Rotate a refresh token",
        "operationId": "postApiAuthRefresh",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Auth"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Refresh"
              }
            }
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "Revoke the current session",
        "operationId": "postApiAuthLogout",
        "tags": [
          "Auth"
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/me": {
      "get": {
        "summary": "Read current account",
        "operationId": "getApiUsersMe",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/products": {
      "get": {
        "summary": "Search active products",
        "operationId": "getApiProducts",
        "tags": [
          "Products"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "enum": [
                "newest",
                "priceAsc",
                "priceDesc",
                "name"
              ]
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "in": "query",
            "name": "size",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 48
            }
          }
        ]
      }
    },
    "/api/products/{id}": {
      "get": {
        "summary": "Read active product",
        "operationId": "getApiProductsId",
        "tags": [
          "Products"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        ]
      }
    },
    "/api/categories": {
      "get": {
        "summary": "List active product categories",
        "operationId": "getApiCategories",
        "tags": [
          "Categories"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/cart": {
      "get": {
        "summary": "Read own cart",
        "operationId": "getApiCart",
        "tags": [
          "Cart"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Clear own cart",
        "operationId": "deleteApiCart",
        "tags": [
          "Cart"
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/cart/items": {
      "post": {
        "summary": "Add quantity to a product in own cart",
        "operationId": "postApiCartItems",
        "tags": [
          "Cart"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or expired"
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "404": {
            "description": "Resource missing or not owned by the caller"
          },
          "409": {
            "description": "Conflict: duplicate value, unavailable stock, invalid transition, reused checkout key, or changed price",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CartInput"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "message"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          }
        }
      },
      "Auth": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "15-minute bearer access token"
          },
          "refreshToken": {
            "type": "string",
            "description": "Rotating refresh token. Never log this value."
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "Register": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 255
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 72,
            "writeOnly": true
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "Login": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "Refresh": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "maxLength": 200,
            "writeOnly": true
          }
        },
        "required": [
          "refreshToken"
        ]
      },
...
      "ProductPage": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "page": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          },
          "totalElements": {
            "type": "integer"
          }
        }
      },
...
    }
  }
}